Move Parameter Stuff Out of Regional MOm6
